# executor: support converting from vectorized evaluation to row-based evaluation