--- a/expression/builtin_vectorized.go
+++ b/expression/builtin_vectorized.go
@@ -103,6 +103,103 @@ func (r *localSliceBuffer) put(buf *chunk.Column) {
 
 type vecRowConverter struct {
 	builtinFunc
+
+	// fields for converting vectorized evaluation to row-based evaluation.
+	buf *chunk.Column
+	sel []int
+}
+
+func newVecRowConverter(underlying builtinFunc) *vecRowConverter {
+	return &vecRowConverter{underlying, nil, nil}
+}
+
+func (c *vecRowConverter) vecEvalRow(evalType types.EvalType, row chunk.Row) (err error) {
+	if c.sel == nil {
+		c.sel = make([]int, 1)
+	}
+	c.sel[0] = row.Idx()
+	if c.buf == nil {
+		c.buf, err = c.builtinFunc.get(evalType, 1)
+		if err != nil {
+			return
+		}
+	}
+	input := row.Chunk()
+	sel := input.Sel()
+	input.SetSel(c.sel)
+	err = c.builtinFunc.vecEval(input, c.buf)
+	input.SetSel(sel)
+	return
+}
+
+func (c *vecRowConverter) evalInt(row chunk.Row) (val int64, isNull bool, err error) {
+	if !c.builtinFunc.vectorized() {
+		return c.builtinFunc.evalInt(row)
+	}
+	if err = c.vecEvalRow(types.ETInt, row); err != nil {
+		return
+	}
+	return c.buf.GetInt64(0), c.buf.IsNull(0), nil
+}
+
+func (c *vecRowConverter) evalReal(row chunk.Row) (val float64, isNull bool, err error) {
+	if !c.builtinFunc.vectorized() {
+		return c.builtinFunc.evalReal(row)
+	}
+	if err = c.vecEvalRow(types.ETReal, row); err != nil {
+		return
+	}
+	return c.buf.GetFloat64(0), c.buf.IsNull(0), nil
+}
+
+func (c *vecRowConverter) evalString(row chunk.Row) (val string, isNull bool, err error) {
+	if !c.builtinFunc.vectorized() {
+		return c.builtinFunc.evalString(row)
+	}
+	if err = c.vecEvalRow(types.ETString, row); err != nil {
+		return
+	}
+	return c.buf.GetString(0), c.buf.IsNull(0), nil
+}
+
+func (c *vecRowConverter) evalDecimal(row chunk.Row) (val *types.MyDecimal, isNull bool, err error) {
+	if !c.builtinFunc.vectorized() {
+		return c.builtinFunc.evalDecimal(row)
+	}
+	if err = c.vecEvalRow(types.ETDecimal, row); err != nil {
+		return
+	}
+	return c.buf.GetDecimal(0), c.buf.IsNull(0), nil
+}
+
+func (c *vecRowConverter) evalTime(row chunk.Row) (val types.Time, isNull bool, err error) {
+	if !c.builtinFunc.vectorized() {
+		return c.builtinFunc.evalTime(row)
+	}
+	if err = c.vecEvalRow(types.ETDatetime, row); err != nil {
+		return
+	}
+	return c.buf.GetTime(0), c.buf.IsNull(0), nil
+}
+
+func (c *vecRowConverter) evalDuration(row chunk.Row) (val types.Duration, isNull bool, err error) {
+	if !c.builtinFunc.vectorized() {
+		return c.builtinFunc.evalDuration(row)
+	}
+	if err = c.vecEvalRow(types.ETReal, row); err != nil {
+		return
+	}
+	return c.buf.GetDuration(0, 0), c.buf.IsNull(0), nil
+}
+
+func (c *vecRowConverter) evalJSON(row chunk.Row) (val json.BinaryJSON, isNull bool, err error) {
+	if !c.builtinFunc.vectorized() {
+		return c.builtinFunc.evalJSON(row)
+	}
+	if err = c.vecEvalRow(types.ETReal, row); err != nil {
+		return
+	}
+	return c.buf.GetJSON(0), c.buf.IsNull(0), nil
 }
 
 func (c *vecRowConverter) vecEval(input *chunk.Chunk, result *chunk.Column) error {
@@ -221,7 +318,7 @@ func (c *vecRowConverter) equal(bf builtinFunc) bool {
 }
 
 func (c *vecRowConverter) Clone() builtinFunc {
-	return &vecRowConverter{c.builtinFunc.Clone()}
+	return &vecRowConverter{c.builtinFunc.Clone(), nil, nil}
 }
 
 type vecRowConvertFuncClass struct {
@@ -233,5 +330,5 @@ func (c *vecRowConvertFuncClass) getFunction(ctx sessionctx.Context, args []Expr
 	if err != nil {
 		return nil, err
 	}
-	return &vecRowConverter{bf}, nil
+	return newVecRowConverter(bf), nil
 }