<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
Following up #11630, support converting from vectorized evaluation to row-based evaluation.

### What is changed and how it works?
Here are all benchmarks:
```
BenchmarkMockDoubleRow/Int-12             100000             19420 ns/op               0 B/op          0 allocs/op
BenchmarkMockDoubleRow/Real-12            100000             20694 ns/op               0 B/op          0 allocs/op
BenchmarkMockDoubleRow/Decimal-12                  30000             58472 ns/op           49152 B/op       1024 allocs/op
BenchmarkMockDoubleRow/Duration-12                 50000             25042 ns/op               0 B/op          0 allocs/op
BenchmarkMockDoubleRow/String-12                   30000             54641 ns/op            7777 B/op       1024 allocs/op
BenchmarkMockDoubleRow/Datetime-12                 10000            179747 ns/op               1 B/op          0 allocs/op
BenchmarkMockDoubleRow/JSON-12                      1000           2191484 ns/op         1960679 B/op      22537 allocs/op

BenchmarkMockDoubleVec2Row/Int-12                  50000             35709 ns/op               0 B/op          0 allocs/op
BenchmarkMockDoubleVec2Row/Real-12                 50000             37853 ns/op               1 B/op          0 allocs/op
BenchmarkMockDoubleVec2Row/Decimal-12              30000             60536 ns/op               0 B/op          0 allocs/op
BenchmarkMockDoubleVec2Row/Duration-12             50000             37666 ns/op               0 B/op          0 allocs/op
BenchmarkMockDoubleVec2Row/String-12               20000             91115 ns/op               0 B/op          0 allocs/op
BenchmarkMockDoubleVec2Row/Datetime-12              5000            227377 ns/op               3 B/op          0 allocs/op
BenchmarkMockDoubleVec2Row/JSON-12                  1000           2286588 ns/op         1959418 B/op      22536 allocs/op
```
There are some performance drawbacks, but it's ok since we will update all expression evaluations which are important to the performance from row-based evaluation to vectorized evaluation soon.

And when all expressions are vectorized, the performance will be much better than now.
Here is the benchmark about pure vectorized evaluation:
```
BenchmarkMockDoubleVec/Int-12            3000000               428 ns/op               0 B/op          0 allocs/op
BenchmarkMockDoubleVec/Real-12           3000000               416 ns/op               0 B/op          0 allocs/op
BenchmarkMockDoubleVec/Decimal-12                 100000             20071 ns/op               0 B/op          0 allocs/op
BenchmarkMockDoubleVec/Duration-12               3000000               605 ns/op               0 B/op          0 allocs/op
BenchmarkMockDoubleVec/String-12                   50000             24638 ns/op               0 B/op          0 allocs/op
BenchmarkMockDoubleVec/Datetime-12                 10000            126060 ns/op               0 B/op          0 allocs/op
BenchmarkMockDoubleVec/JSON-12                      1000           2247046 ns/op         1971854 B/op      22540 allocs/op
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test